+++ src/CosmosDB/CosmosDB/CosmosDBAccount/UpdateAzCosmosDBAccount.cs
@@ -84,6 +84,10 @@ namespace Microsoft.Azure.Commands.CosmosDB
         [Parameter(Mandatory = false, HelpMessage = Constants.DisableKeyBasedMetadataWriteAccessHelpMessage)]
         public bool? DisableKeyBasedMetadataWriteAccess { get; set; }
 
+        [Parameter(Mandatory = false, HelpMessage = Constants.PublicNetworkAccessHelpMessage)]
+        [ValidateNotNullOrEmpty]
+        public string PublicNetworkAccess { get; set; }
+
         [Parameter(Mandatory = false, HelpMessage = Constants.AsJobHelpMessage)]
         public SwitchParameter AsJob { get; set; }
 
@@ -122,6 +126,10 @@ namespace Microsoft.Azure.Commands.CosmosDB
             if (DisableKeyBasedMetadataWriteAccess != null)
             {
                 databaseAccountUpdateParameters.DisableKeyBasedMetadataWriteAccess = DisableKeyBasedMetadataWriteAccess;
+            }
+            if (DisableKeyBasedMetadataWriteAccess != null)
+            {
+                databaseAccountUpdateParameters.PublicNetworkAccess = PublicNetworkAccess;
             }
 
             if (!string.IsNullOrEmpty(DefaultConsistencyLevel))