+++ src/CosmosDB/CosmosDB/Helpers/Constants.cs
@@ -48,6 +48,7 @@ namespace Microsoft.Azure.Commands.CosmosDB.Helpers
         public const string ApiKindHelpMessage = "The type of Cosmos DB database account to create. Accepted values: Sql, MongoDB, Gremlin, Table, Cassandra. Default value: Sql ";
         public const string AccountKeyTypeHelpMessage = "Value from: {ConnectionStrings, Keys, ReadOnlyKeys}. Default is Keys.";
         public const string DisableKeyBasedMetadataWriteAccessHelpMessage = "Disable write operations on metadata resources (databases, containers, throughput) via account keys";
+        public const string PublicNetworkAccessHelpMessage = "Whether or not public endpoint access is allowed for this server. Possible values include: 'Enabled', 'Disabled'";
 
         //Sql cmdlets help messages
         public const string DatabaseNameHelpMessage = "Database name.";